ANDROIDPLOT-85 - Remove all uses of deprecated classes and methods
git-svn-id: file:///Volumes/androidplot-backup/local-svn/ANDROIDPLOT/trunk@72 e6edf6fb-f266-4316-afb4-e53d95876a76

+++ AndroidPlot-Core/src/main/java/com/androidplot/xy/XYStepCalculator.java
@@ -57,7 +57,7 @@ public class XYStepCalculator {
                 break;
             case SUBDIVIDE:
                 stepCount = new Double(stepValue).floatValue();
-                stepPix = (float)(plotPixelSize /(stepCount-1));
+                stepPix = (plotPixelSize /(stepCount-1));
                 stepVal = ValPixConverter.valPerPix(minVal.floatValue(), maxVal.floatValue(), plotPixelSize)*stepPix;
                 //step.setStepCount(stepValue);
                 //step.setStepPix((float)(plotPixelSize/(step.getStepCount()-1)));
